requirements.txt
flask==3.0.0
openai==1.6.1
anthropic>=0.18.0
pandas==2.1.4
scipy==1.11.4
matplotlib==3.8.2
openpyxl==3.1.2
python-dotenv==1.0.0
